renames Groups container from snake to camelcase and updates the import in its parent component
