+++ src/components/Settings.js
@@ -12,7 +12,7 @@ import AccountLinkingContainer from "containers/AccountLinking";
 import SecurityContainer from "containers/Security";
 import DeleteAccount from "containers/DeleteAccount";
 import AccountId from "containers/AccountId";
-import Groups from "../login/components/App/DashboardApp/Settings/Groups/Groups_container";
+import Groups from "../login/components/App/DashboardApp/Settings/Groups/GroupsContainer";
 
 // import "../../node_modules/bootstrap/dist/css/bootstrap.min.css";
 // import "style/base.scss";
